requirements.txt
pyyaml==6.0.1
python-dotenv==1.0.0
openai==1.3.0
opencv-python==4.8.1.78
numpy==1.26.0
pyaudio==0.2.13
baidu-aip==4.16.13
chardet==5.2.0
pvporcupine==3.0.1
psutil==5.9.5 